Switch devmode to UMD format
BREAKING CHANGE:

This allows faster loading in the browser (no need to transpile commonjs
modules).
We give each module a name (equivalent to the original TS sources using
the ///<amd-module name="some/name"/> directive) so that they can be
concatenated together when served to the browser.
Also contains a transformer workaround for an upstream UMD bug in
TypeScript, which is only needed until users move to a TS release that
has the fix.