Implement the optional directoryExists method in our CompilerHost
Only enable this behavior when allowNonHermeticReads is set, since
this option allows TypeScript to go looking around the disk which is
slow on networked file systems.

By implementing the method, we satisfy a check in TypeScript's getAutomaticTypeDirectiveNames
https://github.com/Microsoft/TypeScript/blob/d4c037739500d9f8b42753c57806881a7cd1bd2a/src/compiler/moduleNameResolver.ts#L284

So this permits automatic inclusion of global typings from node_modules/@types.

Fixes #31

PiperOrigin-RevId: 172330884