Strict deps plugin: don't crash when requested file not in program
Angular creates a ts.Program with "stub" files - these are the .ngfactory.ts files
among others which are used for code generation.
These files don't actually exist on disk. The ts.CompilerHost is modified in the
Angular compiler to treat these files as if they do exist, to allow module resolution
to work.

Now in https://github.com/angular/angular/pull/20547 I need to update the Angular
compiler to produce the correct AMD module names (similar use case to how we change
commonjs to goog.module internally, though the mechanism is different.)
We need to treat these generated files as if they are part of the program, when we
do this check:
https://github.com/bazelbuild/rules_typescript/blob/b95f566b7ccc88ecee9613dbea46e7d61abc3a38/internal/tsc_wrapped/compiler_host.ts#L231

There are two possible solutions:
- extract a method like isInProgram(path: string) so that Angular can override it
- add the files to the compilationTargetSrc in bazel

I think the second solution is more general, since we have a number of checks that
predicate on whether the file is one of the requested compilation sources.
This means we will have compilationTargetSrc entries that are not actually on disk,
so we can't assume that we can load all of them from the program.

PiperOrigin-RevId: 176541429