Interop ts_library with closure_js_binary
Every the ts_library impl now returns a struct named
"closure_js_library" needed by both closure_js_library and
closure_js_binary. This struct contains the srcs of the ts_library
transpiled to es6 as well as the path prefixes which should be stripped
off of import statements in order to find the imported file within the
file system.