Introduce a default_tsconfig argument to ts_repositories
If set, this will be used as the tsconfig attribute in any ts_library that doesn't specify one.

This lets users follow the Google model more easily, where they keep a single tsconfig.json file at the root of their sources, to be shared between the editor and Bazel.

Fixes #93

PiperOrigin-RevId: 178917064