Make worker.ts compatible with protobufjs 6
This is a quick fix to allow tsc_wrapped to work with either
protobufjs version 5 or version 6. This can happen if
there is a node_modules with the newer version setup
at the root of the bazel workspace.

A better fix would be to better isolate the tsc_wrapped runtime
from the workspace being compiled, but this is a bit complicated.