Add support for JS transpilation mode to tsc_wrapped.
This mode makes it possible to process files that originated as JS sources, for instance in cases where tsc is used to transpile ES6+ into ES5.

Adds a flag to bazelOpts (isJsTranspilation), which is propagated to tsickle (see usage in https://github.com/angular/tsickle/pull/787). In the context of tsc_wrapped, this flag prevents some not entirely relevant checks from being run on non-TS sources.

PiperOrigin-RevId: 192745472