fix: do not return empty struct if there is no compile action
Currently if there is no action for a `ts_library` that just declares
`.d.ts` files, the `replay_params` will be set to an empty struct.
Design-wise this does not make any sense because there is nothing to be
"replayed".

This fixes an issue where the Angular Bazel rules throw an exception if
there are "replay_params", but are referring to a struct that does not
include the "outputs" property.

```
  target.typescript.replay_params.outputs
struct' object has no attribute 'outputs'
```