Do not depend on the go rules
This PR introduces a `:release` rule which packages `rules_typescript`.
It uses `npm_package` to filter and transform the files in a way
suitable for distribution.

The major improvement of using the artifacts produced by `:release` is
that the users of `rules_typescript` no longer have to build the
`devserver` from source and thus compile the go compiler. Now we don't
depend on `rules_go`, nor `gazelle`.

Fix #353