//internal:tsc_wrapped: fixes for --strictBindCallApply.
at head, :tsc_wrapped is compiled against typescript ~2.9.1 with the
--strict flag. attempting to use this target from a downstream repo
that defines a newer version of typescript fails due to recent
improvements in typescript's type-checking.

specifically, as of typescript 3.2, --strict mode implies
--strictBindCallApply, which brings better type-checking to bind,
call, and apply. this commit fixes a couple of call sites in worker.ts
that fail to compile with --strictBindCallApply.

this commit does not change the version of typescript used to compile
:tsc_wrapped, but is forward-compatible with 3.2 so that downstream
repos can use a later version.