Fix ts_library to produce files that rollup knows how to find.
Currently we produce .closure.js files with ESModules, this naming convention was chosen arbitrarily to indicate we expected these to be inputs to closure compiler.

However other bundlers don't understand this convention and cannot find their inputs. Today we work around this by copying a bunch of files around before running rollup, but this is a performance issue we can avoid just by naming our outputs in the standard way for esmodules.

PiperOrigin-RevId: 269585536