Add the possibility to set a rule name in a conformanceRule's config.
Tsetse's error whitelisting scheme, when used with Bazel, relies on the rule name as the key to the whitelist object, but our current autogenerated rule names for ConformancePattern rules won't cut it: there could be two different BANNED_NAME rules enabled in the same codebase, who'd have colliding names with the current naming rules. With this change, the configuration of a rule lets you specify its rule name to avoid that collision.

PiperOrigin-RevId: 272206672