Fix worker loading when using require.resolve
The breakage was observed downstream in @bazel/worker package which no longer worked.

g3 instructions: There is a copybara transform in third_party/bazel_rules/rules_typescript/copy.bara.sky
which already transforms the protoPath variable to the prefix '../..'
this needs to be changed to '.' so that the output of copybara looks like this PR.
See the README.google.md in the rules_typescript directory for hints about checking that.