construct plugins just before using them
The common plugins includes the tsetse plugin, which grabs
the Program's TypeChecker as part of its construction, which
causes TS to do a bunch of work.

By moving the plugin construction here, that work happens in
the code above (where we explicitly ask TS for diagnostics),
which makes perf trace accounting for type checking more
accurate.

PiperOrigin-RevId: 295773738