Allow all files contained in allowed directory for strictDeps
Currently, if you pass a directory without explicitly
enumerating all files in the directory, the strictDeps rule
will fail. This checks the path of any directories that are
allowed to see if a given files is contained to determine if
it really is transitive or not.

In essence, allow dir/** pattern for strictDeps, as in the case
where a TreeArtifact is used in the srcs of a `ts_library`