Turn back on TypeScript error truncation.
TypeScript runs out of memory trying to produce a human readable
diagnostic in certain situations (infinitely recursing through anonymous
type), see https://github.com/Microsoft/TypeScript/issues/37230.

That's a bug, but more generally speaking, the default TypeScript
behaviour is to run with error truncation. That's for a reason: some
errors are much more readable with truncation. TypeScript's core team
made the call that error truncation is generally a good idea, so we
should follow along unless we have strong data indication otherwise
(which we don't).

Beyond that, error truncation being the default means it is the better
tested path, so we'll probably save us some trouble in the future by
keeping it enabled.

PiperOrigin-RevId: 299096788